requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.1
tzdata==2024.2
asgiref==3.8.1
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
Django==5.1.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1